fix delete row dropdown, 'отмена' close dropdown
