--- src/features/pageAdmin/components/OrderTable/DropdownDeleteRow/DropdownDeleteRow.module.css
@@ -1,8 +1,0 @@
-.overlay {
-  left: 0;
-  bottom: 1.75rem;
-  width: 10rem;
-  padding: 1rem;
-  gap: 0.5rem;
-  box-sizing: border-box;
-}
