+++ src/features/pageAdmin/components/OrderTable/OrderTableDeleteRowDropdown/OrderTableDeleteRowDropdown.module.css
@@ -1,0 +1,20 @@
+._ {
+  position: relative;
+}
+
+.overlay {
+  position: absolute;
+  z-index: 1;
+  display: flex;
+  flex-direction: column;
+  border-radius: 0.5rem;
+  background-color: var(--color-white);
+  left: 0;
+  bottom: 1.75rem;
+  width: 10rem;
+  padding: 1rem;
+  gap: 0.5rem;
+  box-sizing: border-box;
+  box-shadow: 0px 8px 24px rgba(0, 0, 0, 0.1), 0px 4px 8px rgba(0, 0, 0, 0.1),
+    0px 2px 2px rgba(0, 0, 0, 0.1);
+}
